requirements.txt
streamlit
numpy
opencv-python-headless
Pillow
rembg
onnxruntime